requirements.txt
PyMuPDF
pandas
numpy
faiss-cpu
sentence-transformers
torch
openai
tqdm
